requirements.txt
PyYAML==6.0.1
schedule==1.2.1 